Derive PartialEq/Eq to gain StructuralPartialEq
When manually deriving PartialEq one loses the automatic impl of
StructuralPartialEq which is required to for example match on the type.

The problem is that since this type is generic, the TypedUuidKind also
needs to implement PartialEq. But this should not be a problem, as they
are meant to be marker types anyway.

Signed-off-by: Marcel Müller <neikos@neikos.email>